scanners unification plain/simple for reuse code
