+++ ydb/core/tx/columnshard/engines/reader/common_reader/iterator/context.cpp
@@ -9,15 +9,15 @@ namespace NKikimr::NOlap::NReader::NCommon {
 
 TSpecialReadContext::TSpecialReadContext(const std::shared_ptr<TReadContext>& commonContext)
     : CommonContext(commonContext) {
-    auto readMetadata = CommonContext->GetReadMetadataPtrVerifiedAs<TReadMetadata>();
-    Y_ABORT_UNLESS(readMetadata->SelectInfo);
+    ReadMetadata = CommonContext->GetReadMetadataPtrVerifiedAs<TReadMetadata>();
+    Y_ABORT_UNLESS(ReadMetadata->SelectInfo);
 
     double kffAccessors = 0.01;
     double kffFilter = 0.45;
     double kffFetching = 0.45;
     double kffMerge = 0.10;
     TString stagePrefix;
-    if (readMetadata->GetEarlyFilterColumnIds().size()) {
+    if (ReadMetadata->GetEarlyFilterColumnIds().size()) {
         stagePrefix = "EF";
         kffFilter = 0.7;
         kffFetching = 0.15;
@@ -41,15 +41,15 @@ TSpecialReadContext::TSpecialReadContext(const std::shared_ptr<TReadContext>& co
         NGroupedMemoryManager::TScanMemoryLimiterOperator::BuildStageFeatures(stagePrefix + "::MERGE", kffMerge * TGlobalLimits::ScanMemoryLimit)
     };
     ProcessMemoryGuard =
-        NGroupedMemoryManager::TScanMemoryLimiterOperator::BuildProcessGuard(CommonContext->GetReadMetadata()->GetTxId(), stages);
-    ProcessScopeGuard = NGroupedMemoryManager::TScanMemoryLimiterOperator::BuildScopeGuard(
-        CommonContext->GetReadMetadata()->GetTxId(), GetCommonContext()->GetScanId());
+        NGroupedMemoryManager::TScanMemoryLimiterOperator::BuildProcessGuard(ReadMetadata->GetTxId(), stages);
+    ProcessScopeGuard =
+        NGroupedMemoryManager::TScanMemoryLimiterOperator::BuildScopeGuard(ReadMetadata->GetTxId(), GetCommonContext()->GetScanId());
 
-    auto readSchema = readMetadata->GetResultSchema();
+    auto readSchema = ReadMetadata->GetResultSchema();
     SpecColumns = std::make_shared<TColumnsSet>(TIndexInfo::GetSnapshotColumnIdsSet(), readSchema);
-    IndexChecker = readMetadata->GetProgram().GetIndexChecker();
+    IndexChecker = ReadMetadata->GetProgram().GetIndexChecker();
     {
-        auto predicateColumns = readMetadata->GetPKRangesFilter().GetColumnIds(readMetadata->GetIndexInfo());
+        auto predicateColumns = ReadMetadata->GetPKRangesFilter().GetColumnIds(ReadMetadata->GetIndexInfo());
         if (predicateColumns.size()) {
             PredicateColumns = std::make_shared<TColumnsSet>(predicateColumns, readSchema);
         } else {
@@ -58,26 +58,26 @@ TSpecialReadContext::TSpecialReadContext(const std::shared_ptr<TReadContext>& co
     }
     {
         std::set<ui32> columnIds = { NPortion::TSpecialColumns::SPEC_COL_DELETE_FLAG_INDEX };
-        DeletionColumns = std::make_shared<TColumnsSet>(columnIds, readMetadata->GetResultSchema());
+        DeletionColumns = std::make_shared<TColumnsSet>(columnIds, ReadMetadata->GetResultSchema());
     }
 
-    if (!!readMetadata->GetRequestShardingInfo()) {
+    if (!!ReadMetadata->GetRequestShardingInfo()) {
         auto shardingColumnIds =
-            readMetadata->GetIndexInfo().GetColumnIdsVerified(readMetadata->GetRequestShardingInfo()->GetShardingInfo()->GetColumnNames());
-        ShardingColumns = std::make_shared<TColumnsSet>(shardingColumnIds, readMetadata->GetResultSchema());
+            ReadMetadata->GetIndexInfo().GetColumnIdsVerified(ReadMetadata->GetRequestShardingInfo()->GetShardingInfo()->GetColumnNames());
+        ShardingColumns = std::make_shared<TColumnsSet>(shardingColumnIds, ReadMetadata->GetResultSchema());
     } else {
         ShardingColumns = std::make_shared<TColumnsSet>();
     }
     {
-        auto efColumns = readMetadata->GetEarlyFilterColumnIds();
+        auto efColumns = ReadMetadata->GetEarlyFilterColumnIds();
         if (efColumns.size()) {
             EFColumns = std::make_shared<TColumnsSet>(efColumns, readSchema);
         } else {
             EFColumns = std::make_shared<TColumnsSet>();
         }
     }
-    if (readMetadata->HasProcessingColumnIds()) {
-        FFColumns = std::make_shared<TColumnsSet>(readMetadata->GetProcessingColumnIds(), readSchema);
+    if (ReadMetadata->HasProcessingColumnIds()) {
+        FFColumns = std::make_shared<TColumnsSet>(ReadMetadata->GetProcessingColumnIds(), readSchema);
         if (SpecColumns->Contains(*FFColumns) && !EFColumns->IsEmpty()) {
             FFColumns = std::make_shared<TColumnsSet>(*EFColumns + *SpecColumns);
             AFL_DEBUG(NKikimrServices::TX_COLUMNSHARD_SCAN)("ff_modified", FFColumns->DebugString());
@@ -95,7 +95,7 @@ TSpecialReadContext::TSpecialReadContext(const std::shared_ptr<TReadContext>& co
     }
     AllUsageColumns = std::make_shared<TColumnsSet>(*FFColumns + *PredicateColumns);
 
-    PKColumns = std::make_shared<TColumnsSet>(readMetadata->GetPKColumnIds(), readSchema);
+    PKColumns = std::make_shared<TColumnsSet>(ReadMetadata->GetPKColumnIds(), readSchema);
     MergeColumns = std::make_shared<TColumnsSet>(*PKColumns + *SpecColumns);
 
     AFL_DEBUG(NKikimrServices::TX_COLUMNSHARD_SCAN)("columns_context_info", DebugString());
