+++ ydb/core/tx/columnshard/engines/reader/common_reader/iterator/context.h
@@ -2,10 +2,14 @@
 #include "columns_set.h"
 
 #include <ydb/core/tx/columnshard/engines/reader/abstract/read_context.h>
+#include <ydb/core/tx/columnshard/engines/reader/common_reader/constructor/read_metadata.h>
 #include <ydb/core/tx/columnshard/engines/scheme/versions/abstract_scheme.h>
 #include <ydb/core/tx/limiter/grouped_memory/usage/abstract.h>
 
 namespace NKikimr::NOlap::NReader::NCommon {
+
+class TFetchingScript;
+class IDataSource;
 
 class TSpecialReadContext {
 private:
@@ -28,13 +32,36 @@ private:
     YDB_READONLY_DEF(std::shared_ptr<NGroupedMemoryManager::TStageFeatures>, FilterStageMemory);
     YDB_READONLY_DEF(std::shared_ptr<NGroupedMemoryManager::TStageFeatures>, FetchingStageMemory);
 
+    TReadMetadata::TConstPtr ReadMetadata;
     TAtomic AbortFlag = 0;
+
+    virtual std::shared_ptr<TFetchingScript> DoGetColumnsFetchingPlan(const std::shared_ptr<IDataSource>& source) = 0;
 
 protected:
     NIndexes::TIndexCheckerContainer IndexChecker;
     std::shared_ptr<TColumnsSet> EmptyColumns = std::make_shared<TColumnsSet>();
 
 public:
+    template <class T>
+    std::shared_ptr<TFetchingScript> GetColumnsFetchingPlan(const std::shared_ptr<T>& source) {
+        return GetColumnsFetchingPlan(std::static_pointer_cast<IDataSource>(source));
+    }
+
+    std::shared_ptr<TFetchingScript> GetColumnsFetchingPlan(const std::shared_ptr<IDataSource>& source) {
+        return DoGetColumnsFetchingPlan(source);
+    }
+
+    const TReadMetadata::TConstPtr& GetReadMetadata() const {
+        return ReadMetadata;
+    }
+
+    template <class T>
+    std::shared_ptr<T> GetReadMetadataVerifiedAs() const {
+        auto result = std::dynamic_pointer_cast<T>(ReadMetadata);
+        AFL_VERIFY(!!result);
+        return result;
+    }
+
     ui64 GetProcessMemoryControlId() const {
         AFL_VERIFY(ProcessMemoryGuard);
         return ProcessMemoryGuard->GetProcessId();
