+++ ydb/core/tx/columnshard/engines/reader/common_reader/iterator/source.cpp
@@ -1,0 +1,5 @@
+#include "source.h"
+
+namespace NKikimr::NOlap::NReader::NCommon {
+
+}   // namespace NKikimr::NOlap::NReader::NCommon
