+++ ydb/core/tx/columnshard/engines/reader/plain_reader/iterator/constructor.h
@@ -11,15 +11,16 @@ namespace NKikimr::NOlap::NReader::NPlain {
 class TBlobsFetcherTask: public NBlobOperations::NRead::ITask, public NColumnShard::TMonitoringObjectsCounter<TBlobsFetcherTask> {
 private:
     using TBase = NBlobOperations::NRead::ITask;
-    const std::shared_ptr<IDataSource> Source;
+    const std::shared_ptr<NCommon::IDataSource> Source;
     TFetchingScriptCursor Step;
-    const std::shared_ptr<TSpecialReadContext> Context;
+    const std::shared_ptr<NCommon::TSpecialReadContext> Context;
 
     virtual void DoOnDataReady(const std::shared_ptr<NResourceBroker::NSubscribe::TResourcesGuard>& resourcesGuard) override;
     virtual bool DoOnError(const TString& storageId, const TBlobRange& range, const IBlobsReadingAction::TErrorStatus& status) override;
 public:
-    TBlobsFetcherTask(const std::vector<std::shared_ptr<IBlobsReadingAction>>& readActions, const std::shared_ptr<IDataSource>& sourcePtr,
-        const TFetchingScriptCursor& step, const std::shared_ptr<TSpecialReadContext>& context, const TString& taskCustomer, const TString& externalTaskId)
+    TBlobsFetcherTask(const std::vector<std::shared_ptr<IBlobsReadingAction>>& readActions, const std::shared_ptr<NCommon::IDataSource>& sourcePtr,
+        const TFetchingScriptCursor& step, const std::shared_ptr<NCommon::TSpecialReadContext>& context, const TString& taskCustomer,
+        const TString& externalTaskId)
         : TBase(readActions, taskCustomer, externalTaskId) 
         , Source(sourcePtr)
         , Step(step)
