+++ ydb/core/tx/columnshard/engines/reader/plain_reader/iterator/context.cpp
@@ -1,13 +1,15 @@
 #include "context.h"
 #include "source.h"
 
+#include <ydb/core/tx/columnshard/engines/reader/common_reader/iterator/fetch_steps.h>
+#include <ydb/core/tx/columnshard/engines/reader/common_reader/iterator/fetching.h>
 #include <ydb/core/tx/limiter/grouped_memory/usage/service.h>
 
 namespace NKikimr::NOlap::NReader::NPlain {
 
 std::unique_ptr<NArrow::NMerger::TMergePartialStream> TSpecialReadContext::BuildMerger() const {
-    return std::make_unique<NArrow::NMerger::TMergePartialStream>(
-        ReadMetadata->GetReplaceKey(), GetProgramInputColumns()->GetSchema(), GetCommonContext()->IsReverse(), IIndexInfo::GetSnapshotColumnNames());
+    return std::make_unique<NArrow::NMerger::TMergePartialStream>(GetReadMetadata()->GetReplaceKey(), GetProgramInputColumns()->GetSchema(),
+        GetCommonContext()->IsReverse(), IIndexInfo::GetSnapshotColumnNames());
 }
 
 ui64 TSpecialReadContext::GetMemoryForSources(const THashMap<ui32, std::shared_ptr<IDataSource>>& sources) {
@@ -22,12 +24,13 @@ ui64 TSpecialReadContext::GetMemoryForSources(const THashMap<ui32, std::shared_p
     return result;
 }
 
-std::shared_ptr<TFetchingScript> TSpecialReadContext::GetColumnsFetchingPlan(const std::shared_ptr<IDataSource>& source) {
+std::shared_ptr<TFetchingScript> TSpecialReadContext::DoGetColumnsFetchingPlan(const std::shared_ptr<NCommon::IDataSource>& sourceExt) {
+    auto source = std::static_pointer_cast<IDataSource>(sourceExt);
     if (source->NeedAccessorsFetching()) {
         if (!AskAccumulatorsScript) {
             AskAccumulatorsScript = std::make_shared<TFetchingScript>(*this);
             if (ui64 size = source->PredictAccessorsMemory()) {
-                AskAccumulatorsScript->AddStep<TAllocateMemoryStep>(size, EStageFeaturesIndexes::Accessors);
+                AskAccumulatorsScript->AddStep<NCommon::TAllocateMemoryStep>(size, EStageFeaturesIndexes::Accessors);
             }
             AskAccumulatorsScript->AddStep<TPortionAccessorFetchingStep>();
             AskAccumulatorsScript->AddStep<TDetectInMem>(*GetFFColumns());
@@ -46,12 +49,12 @@ std::shared_ptr<TFetchingScript> TSpecialReadContext::GetColumnsFetchingPlan(con
     }();
     const bool useIndexes = (IndexChecker ? source->HasIndexes(IndexChecker->GetIndexIds()) : false);
     const bool isWholeExclusiveSource = source->GetExclusiveIntervalOnly() && source->IsSourceInMemory();
-    const bool needSnapshots = ReadMetadata->GetRequestSnapshot() < source->GetRecordSnapshotMax() || !isWholeExclusiveSource;
+    const bool needSnapshots = GetReadMetadata()->GetRequestSnapshot() < source->GetRecordSnapshotMax() || !isWholeExclusiveSource;
     const bool hasDeletions = source->GetHasDeletions();
     bool needShardingFilter = false;
-    if (!!ReadMetadata->GetRequestShardingInfo()) {
+    if (!!GetReadMetadata()->GetRequestShardingInfo()) {
         auto ver = source->GetShardingVersionOptional();
-        if (!ver || *ver < ReadMetadata->GetRequestShardingInfo()->GetSnapshotVersion()) {
+        if (!ver || *ver < GetReadMetadata()->GetRequestShardingInfo()->GetSnapshotVersion()) {
             needShardingFilter = true;
         }
     }
@@ -75,73 +78,18 @@ std::shared_ptr<TFetchingScript> TSpecialReadContext::GetColumnsFetchingPlan(con
         } else {
             std::shared_ptr<TFetchingScript> result = std::make_shared<TFetchingScript>(*this);
             result->SetBranchName("FAKE");
-            result->AddStep(std::make_shared<TBuildFakeSpec>(source->GetRecordsCount()));
+            result->AddStep<TBuildFakeSpec>();
             return result;
         }
     }
 }
-
-class TColumnsAccumulator {
-private:
-    TColumnsSetIds FetchingReadyColumns;
-    TColumnsSetIds AssemblerReadyColumns;
-    ISnapshotSchema::TPtr FullSchema;
-    std::shared_ptr<TColumnsSetIds> GuaranteeNotOptional;
-
-public:
-    TColumnsAccumulator(const std::shared_ptr<TColumnsSetIds>& guaranteeNotOptional, const ISnapshotSchema::TPtr& fullSchema)
-        : FullSchema(fullSchema)
-        , GuaranteeNotOptional(guaranteeNotOptional) {
-    }
-
-    TColumnsSetIds GetNotFetchedAlready(const TColumnsSetIds& columns) const {
-        return columns - FetchingReadyColumns;
-    }
-
-    bool AddFetchingStep(TFetchingScript& script, const TColumnsSetIds& columns, const EStageFeaturesIndexes stage) {
-        auto actualColumns = GetNotFetchedAlready(columns);
-        FetchingReadyColumns = FetchingReadyColumns + (TColumnsSetIds)columns;
-        if (!actualColumns.IsEmpty()) {
-            script.Allocation(columns.GetColumnIds(), stage, EMemType::Blob);
-            script.AddStep(std::make_shared<TColumnBlobsFetchingStep>(actualColumns));
-            return true;
-        }
-        return false;
-    }
-    bool AddAssembleStep(TFetchingScript& script, const TColumnsSetIds& columns, const TString& purposeId, const EStageFeaturesIndexes stage,
-        const bool sequential) {
-        auto actualColumns = columns - AssemblerReadyColumns;
-        AssemblerReadyColumns = AssemblerReadyColumns + columns;
-        if (actualColumns.IsEmpty()) {
-            return false;
-        }
-        auto actualSet = std::make_shared<TColumnsSet>(actualColumns.GetColumnIds(), FullSchema);
-        if (sequential) {
-            const auto notSequentialColumnIds = GuaranteeNotOptional->Intersect(*actualSet);
-            if (notSequentialColumnIds.size()) {
-                script.Allocation(notSequentialColumnIds, stage, EMemType::Raw);
-                std::shared_ptr<TColumnsSet> cross = actualSet->BuildSamePtr(notSequentialColumnIds);
-                script.AddStep<TAssemblerStep>(cross, purposeId);
-                *actualSet = *actualSet - *cross;
-            }
-            if (!actualSet->IsEmpty()) {
-                script.Allocation(notSequentialColumnIds, stage, EMemType::RawSequential);
-                script.AddStep<TOptionalAssemblerStep>(actualSet, purposeId);
-            }
-        } else {
-            script.Allocation(actualColumns.GetColumnIds(), stage, EMemType::Raw);
-            script.AddStep<TAssemblerStep>(actualSet, purposeId);
-        }
-        return true;
-    }
-};
 
 std::shared_ptr<TFetchingScript> TSpecialReadContext::BuildColumnsFetchingPlan(const bool needSnapshots, const bool exclusiveSource,
     const bool partialUsageByPredicateExt, const bool useIndexes, const bool needFilterSharding, const bool needFilterDeletion) const {
     std::shared_ptr<TFetchingScript> result = std::make_shared<TFetchingScript>(*this);
     const bool partialUsageByPredicate = partialUsageByPredicateExt && GetPredicateColumns()->GetColumnsCount();
 
-    TColumnsAccumulator acc(GetMergeColumns(), ReadMetadata->GetResultSchema());
+    NCommon::TColumnsAccumulator acc(GetMergeColumns(), GetReadMetadata()->GetResultSchema());
     if (!!IndexChecker && useIndexes && exclusiveSource) {
         result->AddStep(std::make_shared<TIndexBlobsFetchingStep>(std::make_shared<TIndexesSet>(IndexChecker->GetIndexIds())));
         result->AddStep(std::make_shared<TApplyIndexStep>(IndexChecker));
@@ -220,11 +168,11 @@ std::shared_ptr<TFetchingScript> TSpecialReadContext::BuildColumnsFetchingPlan(c
             acc.AddAssembleStep(*result, *GetSpecColumns(), "SPEC", EStageFeaturesIndexes::Filter, false);
             result->AddStep(std::make_shared<TSnapshotFilter>());
         }
-        for (auto&& i : ReadMetadata->GetProgram().GetSteps()) {
+        for (auto&& i : GetReadMetadata()->GetProgram().GetSteps()) {
             if (i->GetFilterOriginalColumnIds().empty()) {
                 break;
             }
-            TColumnsSet stepColumnIds(i->GetFilterOriginalColumnIds(), ReadMetadata->GetResultSchema());
+            TColumnsSet stepColumnIds(i->GetFilterOriginalColumnIds(), GetReadMetadata()->GetResultSchema());
             acc.AddAssembleStep(*result, stepColumnIds, "EF", EStageFeaturesIndexes::Filter, false);
             result->AddStep(std::make_shared<TFilterProgramStep>(i));
             if (!i->IsFilterOnly()) {
@@ -257,11 +205,11 @@ std::shared_ptr<TFetchingScript> TSpecialReadContext::BuildColumnsFetchingPlan(c
         if (partialUsageByPredicate) {
             result->AddStep(std::make_shared<TPredicateFilter>());
         }
-        for (auto&& i : ReadMetadata->GetProgram().GetSteps()) {
+        for (auto&& i : GetReadMetadata()->GetProgram().GetSteps()) {
             if (i->GetFilterOriginalColumnIds().empty()) {
                 break;
             }
-            TColumnsSet stepColumnIds(i->GetFilterOriginalColumnIds(), ReadMetadata->GetResultSchema());
+            TColumnsSet stepColumnIds(i->GetFilterOriginalColumnIds(), GetReadMetadata()->GetResultSchema());
             acc.AddAssembleStep(*result, stepColumnIds, "EF", EStageFeaturesIndexes::Filter, false);
             result->AddStep(std::make_shared<TFilterProgramStep>(i));
             if (!i->IsFilterOnly()) {
@@ -271,12 +219,12 @@ std::shared_ptr<TFetchingScript> TSpecialReadContext::BuildColumnsFetchingPlan(c
         acc.AddFetchingStep(*result, *GetFFColumns(), EStageFeaturesIndexes::Fetching);
         acc.AddAssembleStep(*result, *GetFFColumns(), "LAST", EStageFeaturesIndexes::Fetching, !exclusiveSource);
     }
+    result->AddStep<NCommon::TBuildStageResultStep>();
     return result;
 }
 
 TSpecialReadContext::TSpecialReadContext(const std::shared_ptr<TReadContext>& commonContext)
     : TBase(commonContext) {
-    ReadMetadata = GetCommonContext()->GetReadMetadataPtrVerifiedAs<TReadMetadata>();
 }
 
 TString TSpecialReadContext::ProfileDebugString() const {
