+++ ydb/core/tx/columnshard/engines/reader/plain_reader/iterator/context.h
@@ -5,6 +5,7 @@
 #include <ydb/core/tx/columnshard/common/limits.h>
 #include <ydb/core/tx/columnshard/engines/reader/abstract/read_context.h>
 #include <ydb/core/tx/columnshard/engines/reader/common_reader/iterator/context.h>
+#include <ydb/core/tx/columnshard/engines/reader/common_reader/iterator/fetching.h>
 #include <ydb/core/tx/columnshard/engines/reader/plain_reader/constructor/read_metadata.h>
 #include <ydb/core/tx/columnshard/hooks/abstract/abstract.h>
 
@@ -15,6 +16,7 @@ using TColumnsSet = NCommon::TColumnsSet;
 using EStageFeaturesIndexes = NCommon::EStageFeaturesIndexes;
 using TColumnsSetIds = NCommon::TColumnsSetIds;
 using EMemType = NCommon::EMemType;
+using TFetchingScript = NCommon::TFetchingScript;
 
 class TSpecialReadContext: public NCommon::TSpecialReadContext {
 private:
@@ -23,13 +25,14 @@ private:
     YDB_READONLY_DEF(std::shared_ptr<NGroupedMemoryManager::TStageFeatures>, FilterStageMemory);
     YDB_READONLY_DEF(std::shared_ptr<NGroupedMemoryManager::TStageFeatures>, FetchingStageMemory);
 
-    TReadMetadata::TConstPtr ReadMetadata;
     std::shared_ptr<TFetchingScript> BuildColumnsFetchingPlan(const bool needSnapshotsFilter, const bool exclusiveSource,
         const bool partialUsageByPredicate, const bool useIndexes, const bool needFilterSharding, const bool needFilterDeletion) const;
     TMutex Mutex;
     std::array<std::array<std::array<std::array<std::array<std::array<std::optional<std::shared_ptr<TFetchingScript>>, 2>, 2>, 2>, 2>, 2>, 2>
         CacheFetchingScripts;
     std::shared_ptr<TFetchingScript> AskAccumulatorsScript;
+
+    virtual std::shared_ptr<TFetchingScript> DoGetColumnsFetchingPlan(const std::shared_ptr<NCommon::IDataSource>& source) override;
 
 public:
     const ui64 ReduceMemoryIntervalLimit = NYDBTest::TControllers::GetColumnShardController()->GetReduceMemoryIntervalLimit();
@@ -41,17 +44,11 @@ public:
         return MergeStageMemory->GetFullMemory() + FilterStageMemory->GetFullMemory() + FetchingStageMemory->GetFullMemory();
     }
 
-    const TReadMetadata::TConstPtr& GetReadMetadata() const {
-        return ReadMetadata;
-    }
-
     std::unique_ptr<NArrow::NMerger::TMergePartialStream> BuildMerger() const;
 
-    TString ProfileDebugString() const;
+    virtual TString ProfileDebugString() const override;
 
     TSpecialReadContext(const std::shared_ptr<TReadContext>& commonContext);
-
-    std::shared_ptr<TFetchingScript> GetColumnsFetchingPlan(const std::shared_ptr<IDataSource>& source);
 };
 
 }   // namespace NKikimr::NOlap::NReader::NPlain
