+++ ydb/core/tx/columnshard/engines/reader/simple_reader/iterator/context.cpp
@@ -1,25 +1,27 @@
 #include "context.h"
 #include "source.h"
 
+#include <ydb/core/tx/columnshard/engines/reader/common_reader/iterator/fetch_steps.h>
 #include <ydb/core/tx/limiter/grouped_memory/usage/service.h>
 
 namespace NKikimr::NOlap::NReader::NSimple {
 
-std::shared_ptr<TFetchingScript> TSpecialReadContext::GetColumnsFetchingPlan(const std::shared_ptr<IDataSource>& source) {
-    const bool needSnapshots = ReadMetadata->GetRequestSnapshot() < source->GetRecordSnapshotMax();
+std::shared_ptr<TFetchingScript> TSpecialReadContext::DoGetColumnsFetchingPlan(const std::shared_ptr<NCommon::IDataSource>& sourceExt) {
+    const auto source = std::static_pointer_cast<IDataSource>(sourceExt);
+    const bool needSnapshots = GetReadMetadata()->GetRequestSnapshot() < source->GetRecordSnapshotMax();
     if (!needSnapshots && GetFFColumns()->GetColumnIds().size() == 1 &&
         GetFFColumns()->GetColumnIds().contains(NOlap::NPortion::TSpecialColumns::SPEC_COL_PLAN_STEP_INDEX)) {
         std::shared_ptr<TFetchingScript> result = std::make_shared<TFetchingScript>(*this);
         source->SetSourceInMemory(true);
         result->SetBranchName("FAKE");
-        result->AddStep(std::make_shared<TBuildFakeSpec>(source->GetRecordsCount()));
+        result->AddStep<TBuildFakeSpec>();
         result->AddStep<TBuildResultStep>(0, source->GetRecordsCount());
         return result;
     }
     if (!source->GetStageData().HasPortionAccessor()) {
         if (!AskAccumulatorsScript) {
             AskAccumulatorsScript = std::make_shared<TFetchingScript>(*this);
-            AskAccumulatorsScript->AddStep<TAllocateMemoryStep>(
+            AskAccumulatorsScript->AddStep<NCommon::TAllocateMemoryStep>(
                 source->PredictAccessorsSize(GetFFColumns()->GetColumnIds()), EStageFeaturesIndexes::Accessors);
             AskAccumulatorsScript->AddStep<TPortionAccessorFetchingStep>();
             AskAccumulatorsScript->AddStep<TDetectInMem>(*GetFFColumns());
@@ -39,9 +41,9 @@ std::shared_ptr<TFetchingScript> TSpecialReadContext::GetColumnsFetchingPlan(con
     const bool useIndexes = (IndexChecker ? source->HasIndexes(IndexChecker->GetIndexIds()) : false);
     const bool hasDeletions = source->GetHasDeletions();
     bool needShardingFilter = false;
-    if (!!ReadMetadata->GetRequestShardingInfo()) {
+    if (!!GetReadMetadata()->GetRequestShardingInfo()) {
         auto ver = source->GetShardingVersionOptional();
-        if (!ver || *ver < ReadMetadata->GetRequestShardingInfo()->GetSnapshotVersion()) {
+        if (!ver || *ver < GetReadMetadata()->GetRequestShardingInfo()->GetSnapshotVersion()) {
             needShardingFilter = true;
         }
     }
@@ -64,67 +66,12 @@ std::shared_ptr<TFetchingScript> TSpecialReadContext::GetColumnsFetchingPlan(con
     }
 }
 
-class TColumnsAccumulator {
-private:
-    TColumnsSetIds FetchingReadyColumns;
-    TColumnsSetIds AssemblerReadyColumns;
-    ISnapshotSchema::TPtr FullSchema;
-    std::shared_ptr<TColumnsSetIds> GuaranteeNotOptional;
-
-public:
-    TColumnsAccumulator(const std::shared_ptr<TColumnsSetIds>& guaranteeNotOptional, const ISnapshotSchema::TPtr& fullSchema)
-        : FullSchema(fullSchema)
-        , GuaranteeNotOptional(guaranteeNotOptional) {
-    }
-
-    TColumnsSetIds GetNotFetchedAlready(const TColumnsSetIds& columns) const {
-        return columns - FetchingReadyColumns;
-    }
-
-    bool AddFetchingStep(TFetchingScript& script, const TColumnsSetIds& columns, const EStageFeaturesIndexes stage) {
-        auto actualColumns = GetNotFetchedAlready(columns);
-        FetchingReadyColumns = FetchingReadyColumns + (TColumnsSetIds)columns;
-        if (!actualColumns.IsEmpty()) {
-            script.Allocation(columns.GetColumnIds(), stage, EMemType::Blob);
-            script.AddStep(std::make_shared<TColumnBlobsFetchingStep>(actualColumns));
-            return true;
-        }
-        return false;
-    }
-    bool AddAssembleStep(TFetchingScript& script, const TColumnsSetIds& columns, const TString& purposeId, const EStageFeaturesIndexes stage,
-        const bool sequential) {
-        auto actualColumns = columns - AssemblerReadyColumns;
-        AssemblerReadyColumns = AssemblerReadyColumns + columns;
-        if (actualColumns.IsEmpty()) {
-            return false;
-        }
-        auto actualSet = std::make_shared<TColumnsSet>(actualColumns.GetColumnIds(), FullSchema);
-        if (sequential) {
-            const auto notSequentialColumnIds = GuaranteeNotOptional->Intersect(*actualSet);
-            if (notSequentialColumnIds.size()) {
-                script.Allocation(notSequentialColumnIds, stage, EMemType::Raw);
-                std::shared_ptr<TColumnsSet> cross = actualSet->BuildSamePtr(notSequentialColumnIds);
-                script.AddStep<TAssemblerStep>(cross, purposeId);
-                *actualSet = *actualSet - *cross;
-            }
-            if (!actualSet->IsEmpty()) {
-                script.Allocation(notSequentialColumnIds, stage, EMemType::RawSequential);
-                script.AddStep<TOptionalAssemblerStep>(actualSet, purposeId);
-            }
-        } else {
-            script.Allocation(actualColumns.GetColumnIds(), stage, EMemType::Raw);
-            script.AddStep<TAssemblerStep>(actualSet, purposeId);
-        }
-        return true;
-    }
-};
-
 std::shared_ptr<TFetchingScript> TSpecialReadContext::BuildColumnsFetchingPlan(const bool needSnapshots, const bool partialUsageByPredicateExt,
     const bool useIndexes, const bool needFilterSharding, const bool needFilterDeletion) const {
     std::shared_ptr<TFetchingScript> result = std::make_shared<TFetchingScript>(*this);
     const bool partialUsageByPredicate = partialUsageByPredicateExt && GetPredicateColumns()->GetColumnsCount();
 
-    TColumnsAccumulator acc(GetMergeColumns(), ReadMetadata->GetResultSchema());
+    NCommon::TColumnsAccumulator acc(GetMergeColumns(), GetReadMetadata()->GetResultSchema());
     if (!!IndexChecker && useIndexes) {
         result->AddStep(std::make_shared<TIndexBlobsFetchingStep>(std::make_shared<TIndexesSet>(IndexChecker->GetIndexIds())));
         result->AddStep(std::make_shared<TApplyIndexStep>(IndexChecker));
@@ -164,11 +111,11 @@ std::shared_ptr<TFetchingScript> TSpecialReadContext::BuildColumnsFetchingPlan(c
             acc.AddAssembleStep(*result, *GetSpecColumns(), "SPEC", EStageFeaturesIndexes::Filter, false);
             result->AddStep(std::make_shared<TSnapshotFilter>());
         }
-        for (auto&& i : ReadMetadata->GetProgram().GetSteps()) {
+        for (auto&& i : GetReadMetadata()->GetProgram().GetSteps()) {
             if (i->GetFilterOriginalColumnIds().empty()) {
                 break;
             }
-            TColumnsSet stepColumnIds(i->GetFilterOriginalColumnIds(), ReadMetadata->GetResultSchema());
+            TColumnsSet stepColumnIds(i->GetFilterOriginalColumnIds(), GetReadMetadata()->GetResultSchema());
             acc.AddAssembleStep(*result, stepColumnIds, "EF", EStageFeaturesIndexes::Filter, false);
             result->AddStep(std::make_shared<TFilterProgramStep>(i));
             if (!i->IsFilterOnly()) {
@@ -181,13 +128,13 @@ std::shared_ptr<TFetchingScript> TSpecialReadContext::BuildColumnsFetchingPlan(c
         acc.AddFetchingStep(*result, *GetFFColumns(), EStageFeaturesIndexes::Fetching);
         acc.AddAssembleStep(*result, *GetFFColumns(), "LAST", EStageFeaturesIndexes::Fetching, false);
     }
+    result->AddStep<NCommon::TBuildStageResultStep>();
     result->AddStep<TPrepareResultStep>();
     return result;
 }
 
 TSpecialReadContext::TSpecialReadContext(const std::shared_ptr<TReadContext>& commonContext)
     : TBase(commonContext) {
-    ReadMetadata = GetCommonContext()->GetReadMetadataPtrVerifiedAs<TReadMetadata>();
 }
 
 TString TSpecialReadContext::ProfileDebugString() const {
