+++ ydb/core/tx/columnshard/engines/reader/simple_reader/iterator/context.h
@@ -15,11 +15,11 @@ using TColumnsSet = NCommon::TColumnsSet;
 using EStageFeaturesIndexes = NCommon::EStageFeaturesIndexes;
 using TColumnsSetIds = NCommon::TColumnsSetIds;
 using EMemType = NCommon::EMemType;
+using TFetchingScript = NCommon::TFetchingScript;
 
 class TSpecialReadContext: public NCommon::TSpecialReadContext {
 private:
     using TBase = NCommon::TSpecialReadContext;
-    TReadMetadata::TConstPtr ReadMetadata;
     std::shared_ptr<TFetchingScript> BuildColumnsFetchingPlan(const bool needSnapshots, const bool partialUsageByPredicateExt,
         const bool useIndexes, const bool needFilterSharding, const bool needFilterDeletion) const;
     TMutex Mutex;
@@ -27,16 +27,12 @@ private:
         CacheFetchingScripts;
     std::shared_ptr<TFetchingScript> AskAccumulatorsScript;
 
-public:
-    const TReadMetadata::TConstPtr& GetReadMetadata() const {
-        return ReadMetadata;
-    }
+    virtual std::shared_ptr<TFetchingScript> DoGetColumnsFetchingPlan(const std::shared_ptr<NCommon::IDataSource>& source) override;
 
+public:
     virtual TString ProfileDebugString() const override;
 
     TSpecialReadContext(const std::shared_ptr<TReadContext>& commonContext);
-
-    std::shared_ptr<TFetchingScript> GetColumnsFetchingPlan(const std::shared_ptr<IDataSource>& source);
 };
 
 }   // namespace NKikimr::NOlap::NReader::NSimple
