+++ ydb/core/tx/columnshard/engines/reader/simple_reader/iterator/fetching.cpp
@@ -12,67 +12,9 @@
 
 namespace NKikimr::NOlap::NReader::NSimple {
 
-bool TStepAction::DoApply(IDataReader& owner) const {
-    if (FinishedFlag) {
-        AFL_DEBUG(NKikimrServices::TX_COLUMNSHARD_SCAN)("event", "apply");
-        auto* plainReader = static_cast<TPlainReadData*>(&owner);
-        plainReader->MutableScanner().OnSourceReady(Source, nullptr, 0, Source->GetRecordsCount(), *plainReader);
-    }
-    return true;
-}
-
-TConclusionStatus TStepAction::DoExecuteImpl() {
-    if (Source->GetContext()->IsAborted()) {
-        return TConclusionStatus::Success();
-    }
-    auto executeResult = Cursor.Execute(Source);
-    if (!executeResult) {
-        return executeResult;
-    }
-    if (*executeResult) {
-        FinishedFlag = true;
-    }
-    return TConclusionStatus::Success();
-}
-
-TStepAction::TStepAction(const std::shared_ptr<IDataSource>& source, TFetchingScriptCursor&& cursor, const NActors::TActorId& ownerActorId)
-    : TBase(ownerActorId)
-    , Source(source)
-    , Cursor(std::move(cursor))
-    , CountersGuard(Source->GetContext()->GetCommonContext()->GetCounters().GetAssembleTasksGuard()) {
-}
-
-TConclusion<bool> TColumnBlobsFetchingStep::DoExecuteInplace(
-    const std::shared_ptr<IDataSource>& source, const TFetchingScriptCursor& step) const {
-    return !source->StartFetchingColumns(source, step, Columns);
-}
-
-ui64 TColumnBlobsFetchingStep::GetProcessingDataSize(const std::shared_ptr<IDataSource>& source) const {
-    return source->GetColumnBlobBytes(Columns.GetColumnIds());
-}
-
 TConclusion<bool> TIndexBlobsFetchingStep::DoExecuteInplace(
     const std::shared_ptr<IDataSource>& source, const TFetchingScriptCursor& step) const {
     return !source->StartFetchingIndexes(source, step, Indexes);
-}
-
-TConclusion<bool> TAssemblerStep::DoExecuteInplace(const std::shared_ptr<IDataSource>& source, const TFetchingScriptCursor& /*step*/) const {
-    source->AssembleColumns(Columns);
-    return true;
-}
-
-ui64 TAssemblerStep::GetProcessingDataSize(const std::shared_ptr<IDataSource>& source) const {
-    return source->GetColumnRawBytes(Columns->GetColumnIds());
-}
-
-TConclusion<bool> TOptionalAssemblerStep::DoExecuteInplace(
-    const std::shared_ptr<IDataSource>& source, const TFetchingScriptCursor& /*step*/) const {
-    source->AssembleColumns(Columns, !source->IsSourceInMemory());
-    return true;
-}
-
-ui64 TOptionalAssemblerStep::GetProcessingDataSize(const std::shared_ptr<IDataSource>& source) const {
-    return source->GetColumnsVolume(Columns->GetColumnIds(), EMemType::RawSequential);
 }
 
 TConclusion<bool> TFilterProgramStep::DoExecuteInplace(const std::shared_ptr<IDataSource>& source, const TFetchingScriptCursor& /*step*/) const {
@@ -130,162 +72,9 @@ TConclusion<bool> TShardingFilter::DoExecuteInplace(const std::shared_ptr<IDataS
     return true;
 }
 
-TConclusion<bool> TBuildFakeSpec::DoExecuteInplace(const std::shared_ptr<IDataSource>& source, const TFetchingScriptCursor& /*step*/) const {
-    std::vector<std::shared_ptr<arrow::Array>> columns;
-    for (auto&& f : IIndexInfo::ArrowSchemaSnapshot()->fields()) {
-        columns.emplace_back(NArrow::TThreadSimpleArraysCache::GetConst(f->type(), NArrow::DefaultScalar(f->type()), Count));
-    }
-    source->MutableStageData().AddBatch(
-        std::make_shared<NArrow::TGeneralContainer>(arrow::RecordBatch::Make(TIndexInfo::ArrowSchemaSnapshot(), Count, columns)));
-    source->SetUsedRawBytes(0);
-    source->Finalize({});
-    return true;
-}
-
 TConclusion<bool> TApplyIndexStep::DoExecuteInplace(const std::shared_ptr<IDataSource>& source, const TFetchingScriptCursor& /*step*/) const {
     source->ApplyIndex(IndexChecker);
     return true;
-}
-
-TConclusion<bool> TFetchingScriptCursor::Execute(const std::shared_ptr<IDataSource>& source) {
-    AFL_VERIFY(source);
-    NMiniKQL::TThrowingBindTerminator bind;
-    Script->OnExecute();
-    while (!Script->IsFinished(CurrentStepIdx)) {
-        if (source->HasStageData() && source->GetStageData().IsEmpty()) {
-            source->OnEmptyStageData();
-            break;
-        }
-        auto step = Script->GetStep(CurrentStepIdx);
-        TMemoryProfileGuard mGuard("SCAN_PROFILE::FETCHING::" + step->GetName() + "::" + Script->GetBranchName(),
-            IS_DEBUG_LOG_ENABLED(NKikimrServices::TX_COLUMNSHARD_SCAN_MEMORY));
-        AFL_DEBUG(NKikimrServices::TX_COLUMNSHARD_SCAN)("scan_step", step->DebugString())("scan_step_idx", CurrentStepIdx);
-        AFL_VERIFY(!CurrentStartInstant);
-        CurrentStartInstant = TMonotonic::Now();
-        AFL_VERIFY(!CurrentStartDataSize);
-        CurrentStartDataSize = step->GetProcessingDataSize(source);
-        const TConclusion<bool> resultStep = step->ExecuteInplace(source, *this);
-        if (!resultStep) {
-            return resultStep;
-        }
-        if (!*resultStep) {
-            return false;
-        }
-        FlushDuration();
-        ++CurrentStepIdx;
-    }
-    return true;
-}
-
-bool TAllocateMemoryStep::TFetchingStepAllocation::DoOnAllocated(std::shared_ptr<NGroupedMemoryManager::TAllocationGuard>&& guard,
-    const std::shared_ptr<NGroupedMemoryManager::IAllocation>& /*allocation*/) {
-    auto data = Source.lock();
-    if (!data || data->GetContext()->IsAborted()) {
-        guard->Release();
-        return false;
-    }
-    if (StageIndex == EStageFeaturesIndexes::Accessors) {
-        data->MutableStageData().SetAccessorsGuard(std::move(guard));
-    } else {
-        data->RegisterAllocationGuard(std::move(guard));
-    }
-    Step.Next();
-    auto task = std::make_shared<TStepAction>(data, std::move(Step), data->GetContext()->GetCommonContext()->GetScanActorId());
-    NConveyor::TScanServiceOperator::SendTaskToExecute(task);
-    return true;
-}
-
-TAllocateMemoryStep::TFetchingStepAllocation::TFetchingStepAllocation(
-    const std::shared_ptr<IDataSource>& source, const ui64 mem, const TFetchingScriptCursor& step, const EStageFeaturesIndexes stageIndex)
-    : TBase(mem)
-    , Source(source)
-    , Step(step)
-    , TasksGuard(source->GetContext()->GetCommonContext()->GetCounters().GetResourcesAllocationTasksGuard())
-    , StageIndex(stageIndex) {
-}
-
-void TAllocateMemoryStep::TFetchingStepAllocation::DoOnAllocationImpossible(const TString& errorMessage) {
-    auto sourcePtr = Source.lock();
-    if (sourcePtr) {
-        sourcePtr->GetContext()->GetCommonContext()->AbortWithError(
-            "cannot allocate memory for step " + Step.GetName() + ": '" + errorMessage + "'");
-    }
-}
-
-TConclusion<bool> TAllocateMemoryStep::DoExecuteInplace(const std::shared_ptr<IDataSource>& source, const TFetchingScriptCursor& step) const {
-    ui64 size = PredefinedSize.value_or(0);
-    for (auto&& i : Packs) {
-        ui32 sizeLocal = source->GetColumnsVolume(i.GetColumns().GetColumnIds(), i.GetMemType());
-        if (source->GetStageData().GetUseFilter() && i.GetMemType() != EMemType::Blob && source->GetContext()->GetReadMetadata()->HasLimit()) {
-            const ui32 filtered =
-                source->GetStageData().GetFilteredCount(source->GetRecordsCount(), source->GetContext()->GetReadMetadata()->GetLimitRobust());
-            if (filtered < source->GetRecordsCount()) {
-                sizeLocal = sizeLocal * 1.0 * filtered / source->GetRecordsCount();
-            }
-        }
-        size += sizeLocal;
-    }
-
-    auto allocation = std::make_shared<TFetchingStepAllocation>(source, size, step, StageIndex);
-    NGroupedMemoryManager::TScanMemoryLimiterOperator::SendToAllocation(source->GetContext()->GetProcessMemoryControlId(),
-        source->GetContext()->GetCommonContext()->GetScanId(), source->GetMemoryGroupId(), { allocation }, (ui32)StageIndex);
-    return false;
-}
-
-ui64 TAllocateMemoryStep::GetProcessingDataSize(const std::shared_ptr<IDataSource>& /*source*/) const {
-    return 0;
-}
-
-TString TFetchingScript::DebugString() const {
-    TStringBuilder sb;
-    TStringBuilder sbBranch;
-    for (auto&& i : Steps) {
-        if (i->GetSumDuration() > TDuration::MilliSeconds(10)) {
-            sbBranch << "{" << i->DebugString() << "};";
-        }
-    }
-    if (!sbBranch) {
-        return "";
-    }
-    sb << "{branch:" << BranchName << ";limit:" << Limit << ";";
-    if (FinishInstant && StartInstant) {
-        sb << "duration:" << *FinishInstant - *StartInstant << ";";
-    }
-
-    sb << "steps_10Ms:[" << sbBranch << "]}";
-    return sb;
-}
-
-TFetchingScript::TFetchingScript(const TSpecialReadContext& context)
-    : Limit(context.GetReadMetadata()->GetLimitRobust()) {
-}
-
-void TFetchingScript::Allocation(const std::set<ui32>& entityIds, const EStageFeaturesIndexes stage, const EMemType mType) {
-    if (Steps.size() == 0) {
-        AddStep<TAllocateMemoryStep>(entityIds, mType, stage);
-    } else {
-        std::optional<ui32> addIndex;
-        for (i32 i = Steps.size() - 1; i >= 0; --i) {
-            if (auto allocation = std::dynamic_pointer_cast<TAllocateMemoryStep>(Steps[i])) {
-                if (allocation->GetStage() == stage) {
-                    allocation->AddAllocation(entityIds, mType);
-                    return;
-                } else {
-                    addIndex = i + 1;
-                }
-                break;
-            } else if (std::dynamic_pointer_cast<TAssemblerStep>(Steps[i])) {
-                continue;
-            } else if (std::dynamic_pointer_cast<TColumnBlobsFetchingStep>(Steps[i])) {
-                continue;
-            } else {
-                addIndex = i + 1;
-                break;
-            }
-        }
-        AFL_VERIFY(addIndex);
-        InsertStep<TAllocateMemoryStep>(*addIndex, entityIds, mType, stage);
-    }
 }
 
 NKikimr::TConclusion<bool> TFilterCutLimit::DoExecuteInplace(
@@ -385,7 +174,6 @@ TConclusion<bool> TBuildResultStep::DoExecuteInplace(const std::shared_ptr<IData
 }
 
 TConclusion<bool> TPrepareResultStep::DoExecuteInplace(const std::shared_ptr<IDataSource>& source, const TFetchingScriptCursor& /*step*/) const {
-    source->Finalize(NYDBTest::TControllers::GetColumnShardController()->GetMemoryLimitScanPortion());
     std::shared_ptr<TFetchingScript> plan = std::make_shared<TFetchingScript>(*source->GetContext());
     if (source->IsSourceInMemory()) {
         AFL_VERIFY(source->GetStageResult().GetPagesToResultVerified().size() == 1);
@@ -404,6 +192,18 @@ TConclusion<bool> TPrepareResultStep::DoExecuteInplace(const std::shared_ptr<IDa
     auto task = std::make_shared<TStepAction>(source, std::move(cursor), source->GetContext()->GetCommonContext()->GetScanActorId());
     NConveyor::TScanServiceOperator::SendTaskToExecute(task);
     return false;
+}
+
+TConclusion<bool> TBuildFakeSpec::DoExecuteInplace(const std::shared_ptr<IDataSource>& source, const TFetchingScriptCursor& /*step*/) const {
+    std::vector<std::shared_ptr<arrow::Array>> columns;
+    for (auto&& f : IIndexInfo::ArrowSchemaSnapshot()->fields()) {
+        columns.emplace_back(NArrow::TThreadSimpleArraysCache::GetConst(f->type(), NArrow::DefaultScalar(f->type()), source->GetRecordsCount()));
+    }
+    source->MutableStageData().AddBatch(std::make_shared<NArrow::TGeneralContainer>(
+        arrow::RecordBatch::Make(TIndexInfo::ArrowSchemaSnapshot(), source->GetRecordsCount(), columns)));
+    source->SetUsedRawBytes(0);
+    source->Finalize({});
+    return true;
 }
 
 }   // namespace NKikimr::NOlap::NReader::NSimple
